+++ pheasant/core/renderer.py
@@ -6,14 +6,17 @@ from typing import Any, Dict, List, Optional, Union
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 
 from pheasant.core.base import Base, get_render_name
+from pheasant.core.page import Page
 from pheasant.core.parser import Parser, Render  # Render is type, not class
 
 
 class Renderer(Base):
     patterns: Dict[str, str] = field(default_factory=dict, init=False)
     renders: Dict[str, Render] = field(default_factory=dict, init=False)
+    page: Page = field(default_factory=Page, init=False)
     _parser: Optional[Parser] = field(default=None, init=False)
-    _src_path: str = ""
+
+    class_parser = None
 
     def __post_init__(self):
         super().__post_init__()
@@ -26,8 +29,16 @@ class Renderer(Base):
         """Called from __post_init__."""
         pass
 
-    def reset(self) -> None:
-        """Called per page."""
+    def start(self) -> None:
+        """Called at conversion start"""
+        pass
+
+    def enter(self) -> None:
+        """Called at page enter event"""
+        pass
+
+    def exit(self) -> None:
+        """Called at page exit event"""
         pass
 
     def register(self, pattern: str, render: Render, render_name: str = "") -> None:
@@ -44,18 +55,13 @@ class Renderer(Base):
 
     @parser.setter
     def parser(self, parser: Parser) -> None:
+        self._parser = self.configure_parser(parser)
+
+    def configure_parser(self, parser: Parser) -> Parser:
         for render_name, pattern in self.patterns.items():
             render = self.renders[render_name]
             parser.register(pattern, render, render_name)
-        self._parser = parser
-
-    @property
-    def src_path(self) -> str:
-        return self._src_path
-
-    @src_path.setter
-    def src_path(self, src_path: str) -> None:
-        self._src_path = src_path
+        return parser
 
     def set_template(self, names: Union[str, List[str]], directory: str = ".") -> None:
         module = importlib.import_module(self.__module__)
@@ -75,11 +81,18 @@ class Renderer(Base):
         template = self.config[f"{name}_template"]
         return template.render(context, config=self.config, **kwargs)
 
-    def parse(self, source: str) -> str:
-        return self.parser.parse(source, decorate=self.decorate)
+    def parse(self, source: str = "") -> str:
+        return self.parser.parse(source or self.page.source, decorate=self.decorate)
 
     def decorate(self, cell):
         """Decorate cell.output after parse.
 
         Overwritten by subclass."""
         pass
+
+    def findall(self, source: str = "") -> List:
+        if self.class_parser is None:
+            self.class_parser = Parser()
+            self.configure_parser(self.class_parser)
+
+        return self.class_parser.findall(source or self.page.source)