+++ tests/renderers/jupyter/test_jupyter_renderer.py
@@ -1,23 +1,5 @@
-from datetime import timedelta
-
-from pheasant.renderers.jupyter.jupyter import (format_timedelta,
-                                                replace_for_display)
-
-
-def test_format_timedelta():
-    assert format_timedelta(timedelta(days=1, seconds=100)) == "24h1min40s"
-    assert format_timedelta(timedelta(seconds=3670)) == "1h1min10s"
-    assert format_timedelta(timedelta(seconds=2670)) == "44min30s"
-    assert format_timedelta(timedelta(seconds=59.564)) == "59.6s"
-    assert format_timedelta(timedelta(seconds=9.3)) == "9.30s"
-    assert format_timedelta(timedelta(seconds=0.123)) == "123ms"
-    assert format_timedelta(timedelta(seconds=0.0453)) == "45.3ms"
-    assert format_timedelta(timedelta(seconds=0.006544)) == "6.54ms"
-    assert format_timedelta(timedelta(seconds=3.46e-4)) == "346us"
-    assert format_timedelta(timedelta(seconds=3.42e-5)) == "34us"
-    assert format_timedelta(timedelta(seconds=5.934e-6)) == "6us"
-    assert format_timedelta(timedelta(seconds=5.934e-7)) == "1us"
-    assert format_timedelta(timedelta(seconds=3.934e-7)) == "<1us"
+from pheasant.renderers.jupyter.jupyter import replace_for_display
+from pheasant.renderers.jupyter.display import extra_html
 
 
 def test_render_inline_option(jupyter):
@@ -29,11 +11,11 @@ def test_render_inline_option(jupyter):
     jupyter.execute("plot = figure(plot_width=250, plot_height=250)")
     output = jupyter.parse("```python inline\nplot\n```\n")
     assert '<div class="cell jupyter display"' in output
-    assert "bokeh" in jupyter.meta[""]["extra_module"]
+    assert jupyter.cache[-1].extra_module == "bokeh"
 
     output = jupyter.parse("```python inline\nhv.Curve(([1,2],[3,4]))\n```\n")
     assert '<div class="cell jupyter display"' in output
-    assert "holoviews" in jupyter.meta[""]["extra_module"]
+    assert jupyter.cache[-1].extra_module == "holoviews"
 
     jupyter.execute(
         (
@@ -50,13 +32,18 @@ def test_render_inline_option(jupyter):
         )
     )
     assert "document.addEventListener" in output
-    assert "altair" in jupyter.meta[""]["extra_module"]
+    assert jupyter.cache[-1].extra_module == "altair"
 
     _ = jupyter.parse("```python inline\nplot\n```\n")
 
 
-def test_render_extra_html(jupyter):
-    extra = jupyter.extra_html
+def test_render_extra_modules_and_html(jupyter):
+    extra_modules = set(jupyter.get_extra_modules())
+    assert "bokeh" in extra_modules
+    assert "holoviews" in extra_modules
+    assert "altair" in extra_modules
+
+    extra = extra_html(extra_modules)
     assert "bokeh" in extra
     assert "HoloView" in extra
     assert "vega" in extra
@@ -108,4 +95,4 @@ def test_render_fenced_code_with_option():
 
 def test_render_bare_execute_count(jupyter):
     source = "{{2*3}}\n```python\n1\n```\n{{2}}\n```python\n1\n```\n"
-    assert jupyter.bare_execute_count(source) == 4
+    assert len(jupyter.findall(source)) == 4